+++ Beam/config/google.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    'ga_reporting_enabled' => env('GOOGLE_GA_REPORTING_ENABLED', false),
-    'app_name' => env('GOOGLE_APP_NAME', ''),
-    'service_account_file' => env('GOOGLE_SERVICE_ACCOUNT_FILE', ''),
-    'ga_view_id' => env('GOOGLE_GA_VIEW_ID', ''),
-];