+++ Beam/extensions/beam-module/resources/assets/js/bootstrap.js
@@ -17,7 +17,6 @@ global.DashboardRoot = require("./components/dashboard/DashboardRoot.vue").defau
 global.ArticleDetails = require("./components/dashboard/ArticleDetails.vue").default;
 global.UserPath = require("./components/userpath/UserPath.vue").default;
 global.ConversionsSankeyDiagram = require("./components/userpath/ConversionsSankeyDiagram.vue").default;
-global.GoogleAnalyticsReportingHistogram = require("./components/dashboard/GoogleAnalyticsReportingHistogram.vue").default;
 global.DashboardStore = require("./components/dashboard/store.js").default;
 
 global.SmartRangeSelector = require("@remp/js-commons/js/components/SmartRangeSelector.vue").default;